New method, time domain filtering with block convolution
An alternative method for filtering, in time domain using fast block
convolution instead of freq. domain. Adding a few new input parameters
as well, default is still the same as before. The check for method
selected makes freq. filtering code appear as if it was changed for
github, but NO changes were made there for this update.