MMS Re-compiled irfu_cdfwrite_*.mexa64 functions.
As the SDC runs an older verison (2.11) of libc.so.6 it was required to recompile these files (previously compiled with gcc version 4.7, which used v2.15 of libc.so.6) using an older version of gcc (brain uses libc.so.6 v.2.9). However it may still be some issues with these files as Matlab 2013a (on brain) is recommended to use gcc version 4.4.x and not the brain installed 4.1. If problems occur still try recompiling on SDC directly, by first compiling libcdf.a and then configuring Matlab Mex to use gcc with std=c99 instead of ansi.