MMS Naming of output files and spelling correction of debug message
The MMS CDF filenames should all be lower case, even the optionalDescriptor. And according to e-mail from Mark (UHN) 20140203 the FIELDS group consider SITL to be a dataLevel and as such it determines the name and final location of STIL CDF output files. Also some minor spelling mistakes in debug message and filename are corrected. ( STIL != SITL, received allright?)