MMS Updated MMS to one cdfwrite, help messages, ..
MMS files have been updated to include a short help message/description.
MMS_init will try to create the compiled mexa64 file if it does not exist for the mms_sdc_cdfwrite function. For this to function the mex command is called in bash/unix but for now CFLAG std=c99 must be set manually in mexopt.sh first. This argument should be included in the calling of mex... But it require a single ' which Matlab interpets as beginnig/end of string...

MMS will after this commit start development on a separate branch.