MMS create a .m file interface mms_sdc_cdfwrite
Created a new Matlab interface file for writing to MMS CDF files, this way some help are provided making it a bit easier..
If the mexa64 file does not previously exist it will now be build when trying to write CDF files, instead of directly at mms_init.
A log message is still created notifying user that the mexa64 file was not found but built.

The flip of matricies is performed in the new interface file, that way users will simply send data in the format they expect it to be written,
i.e. for a matrix like xyz_pgse
[x1, y1, z1; x2, y2, z2]; will be written as two records where [x1, y1, z1] is the first record.

Test functions for OnlyWriting and functions used to write data in normal processing has been updated accordingly.