MMS move writing of log file(-s) to subdirectory
It was pointed out that logs should be kept seperate for each spacecraft and instrument processing. To accomodate this all logs are now placed in $LOG_PATH_ROOT / mmsX / sdp /.

The empty logfile for each data output product is also placed in same subfolder. This empty log files name is used by SDC scripts to find the filename of outputs created.