MMS launch scripts for SDC processing added to git
These are the files used for launching the various MMS SDP processing done at SDC from bash.
Important is however: Move these files to the mms-sdp/~software/ folder (which has irfu-matlab as a subdir) on the SDC machines, this way the matlab path is correctly defined (excluding .git folder etc.).