Closes #40 Note: For most of irfu-matlab usage the IRBEM package is NOT required and the warning could be ignored!
I was not able to (easily) resolve the dependencies for the IRBEM, Matlab support no free fortran compiler for Windows and the provided "onera_desp_lib_Win64_x86.dll" library have a lot of dependencies (potentially a full install of MS Visual Studio and its redistributables could be required). Users are encuraged to configure it themselves, if required.