Try multiple times to read ancillary files
For some, yet unknown, reason sometimes the loading of ancillary files fail at SDC. The problem result in "Error MATLAB:textscan:NotAnIntegerValue with message: Header lines must be integer-valued.", and appears to be releated to problem with "grep'ing" for the last comment/header line. This was the case for the two failed processing of data which occurred 2017-11-17T15 CET (processing burst data from MMS1 and MMS2 for the L2Pre dce2d). When I tried to replicate this problem (grepping the files) the resulting nHeaders were correct every single time, therefor the new code simply tries to re-run the grep command (at most 5 times in a row).