MMS new log file names and paths
The old approach with writing logs directly to $LOG_PATH_ROOT/mms[1-4]/edp/ is not good. It was descided that we should move all written logs into subfolders similar to the data files with names based on what process and data being written to the logs. As such we (SDP) have now logs with name format of
mms[1-4]_edp_{tmmode}_{level}_{descriptor}_{dateYYYMMDDhhmmss}_{runTime}.log
put into folders $LOG_PATH_ROOT/mms[1-4]/edp/{tmmode}/{level}/{descriptor}/{YYYY}/{MM}/ (with possible {DD}/ at the end for burst data).

NOTE: For now this somewhat breaks compatibility in that the order of files in input argument now must start with the lowest level data to be processed (L1b in the case of L2a & QL & SCPOT, as well as for L2pre burst processing where the Fast L2a is also used).