MMS DONT MERGE. MMS4p4 fail BAD_BIAS => ASSYM_CONF
As agreed to in MMS SDP meeting 2017/11/21T13 to 15 CET, bitmask "BAD_BIAS" should no longer be used to mark mms4p4 after it's failure. Instead this BAD_BIAS should correspond more to "non-optimal" bias settings, probe failure is now bitmasked by "ASSYM_CONF" in reconstructed/measured DCE/DCV. This "BAD_BIAS" is set solely by bias_check(), as a result the mms_sdp_limit_bias are made a lot more relaxed as to its limits at the end of the interval. This relaxation is to allow for new biases to be commanded without impacting processing. (They could be made more strict for old intervals where the bias will not be commanded to new values, as long as the last segment is relaxed.). One clear example of what will now be marked as "BAD_BIAS" is all of the eclipses (of either the Earth or the Moon) where the SDP probe biases are all set to zero. A second example of moon eclipse: 2017/11/19T1450.

Note: Next up is the change to Quality (and spacecraft potential file), so that "BAD_BIAS" will lower the quality sufficently for the SITL production (of SITL products DCE and SCPOT) to simply place FillValues for intervals with "BAD_BIAS" (where "quality is less or equal to one"???).

(Also included here is some minor clean up.).