MMS Read new CAL files containing "regions" TBD.
As agreed during meeting 2017/11/21T13 to 15 CET, a new calibration file is to be provided by Daniel. This will contain start times and a flag to indicate what the interval following it mainy have for environment. With flag value == 0: default/unknown, flag == 1: Solar Wind.
Values with flag == 1 (ie Solar wind) is to have the solar wind removal code running with relaxed conditions to model to catch as many wakes as possible.