MMS DONT MERGE, BAD_BIAS & Maneuvers => FillValues
That is for QL and SCPOT products, the output DSL electric fields and scpot will be filled with FillValues during intervals where BAD_BIAS or MANEUVERS are identified (typically during eclipse zero current bias is applied). This will force SITL products to not have anything other than FillValue during these times, but for anyone who may be interested the other data will not be touched (probe2sc or individual probes) as well as L2pre E-field products (at least with this commit).
Quality is lowered to zero for all our products during such intervals.

The intervals of eclipse (BAD_BIAS) are not perfectly identified as HK resolution is not sufficient to detect the instance of change bias.

The reason for this commit is that for the second time around, the MMS has been in eclipse from the Moon (now started around 2017/11/19T14.50 UT) and for a second time the SITL person has selected this as an interesting interval based on E-field measurements... (We may find it interesting but not primarily because of the e-field but because of the different photo emission). Previous Moon eclipse was also selected and similarly some Manuevers have been selected, with this change the SITL's will not see our data during these intervals and hopefully not select them anymore!