MMS DONT MERGE, eclipse info from mission timeline
The resolution of HK10e is only 64 seconds meaning it does not capture the start of eclipses well, and sometimes (12 times in the time period 2017/07/01->2017/11/20) the Hk10e indicate a large biased current as it caputred one information during a sweep. Therefore try to use eclipse information which was available in the mission timeline xml files (along with maneuvers).

Note: Changes and clearification in code and help text to follow.

Improves commit 688c6993fa0361f7ba20271794bc79d86ce408d7.