DONT MERGE Outline fix for multidim. cdf variables
This is only an outline of code to fix the problem with multidimensional data being read from CDF files when these have only one record in the main dependency (Depend_0).
When data is read from CDF it is at first loaded as for instance "3x3xN" (where N is the number of records), this is then permutated into "Nx3x3" by our code (with the main Depend_0 / Epoch being an array of length N). A problem was identified by @danbgraham, using mms1_fpi_fast_ql_dis_20171026100000_v3.2.0.cdf, when there is only one record written to a particular CDF file. In that case the data read becomes "3x3x1" which in Matlab is equal to "3x3" and subsequently not identified in our present code.. This outline of code looks at the info returned from spdfcdfinfo instead, but some testing is still to be done.

Note: Testing not yet fully completed!