Speed up ".v" for PDist and correct v's phi
There is no need to loop through all timestamps for the case with coordinates and energies that do not change with time.
Note: Should we not use obj.mass() to ensure it selects ion mass or electron mass depending on obj? And should we not also make sure data is in the expected units?