Why run with multiple interp1 in multiple loops?
Should it not be possible to simply run it with interpn() making it alot quicker? The only reason I can think of, at this moment, is that the default method used ('pchip') only works with interp1(). So therefor I leave it mostly as is, only very slight speed up reducing the number of calls to size().