Add one final License text,
(in order to comply with the license).

By the looks of it Carlos added this license text back in 2016/05/04 (which was after I wrote "license unspecified" in commit 68ec8afbb874cc58498b62047c3fabaf53e85dec). I believe we are now finally fully compliant with all the licenses of all the contributed code segments in irfu-matlab. :)

Also correct .m file permissions to not have the executable bit set.