Move polarplot->polarplot3, keep Matlab built in..
I had some problems creating simple polar plots, it was caused by this polarplot() being called and not the Matlab built in polarplot() function. (It was itroduced in R2016a). I did not locate any calls to polarplot() in our irfu-matlab code, and as we already had a polarplot2(), our old polarplot() is now renamed to polarplot3().