Update IRBEM-LIB, avoid problem with Matlab R2018a
The problem is that the code use hard coded values for what appears to be IGRF (11th ed.), which are out of date. This new code still have hard coded values, but they are updated to IGRF 12th ed. and as such valid until 2020. (It would be alot better if the code relied on an environment variable, which we could point towards our included "igrf12coeffs.txt". The code would then only fall back to the hard coded values if this variable was not set, similar to the leap seconds table for the CDF patch, but sadly it does not appear to have this capability.)

The problem manifested it self as a lot of warnings such as:
 *** WARNING -- Input year = 2018.50
  is out of valid range 1965-2015. Using nearest ***

Source used here was the svn revision 610, dated 2018-01-17T08:55:24 CET, available from: https://sourceforge.net/p/irbem/code/610/tree/

Note 1: The Mac 64 bit version was not updated with this commit and will have to be added separatly. (When I locate a Mac with Fortran installed...)

Note 2: The Windows 32 and 64 bit DLL's are included here but for the 64 bit version the instructions at https://sourceforge.net/p/irbem/code/HEAD/tree/trunk/manual/frames/matlab.html?format=raw would probably still have to be followed (a lot of dependancies).