Manually fix remaining checkcode() in irf_plot.m
As mentioned in commit be0551735e56b7af1998a4a4a23b37f00ca508e0 the sharad variable "tint" must in a future version of Matlab be defined in the parent function if it is to be used by a child function (and its result value be used by the parent afterwards). This is detected by checkcode() as well as warning given when executing irf_plot() but unfortuantly this (and the other changes here) can not be solved automatically by use of checkcode().