Dont use FillValue in scpot re-processing old data
Following e-mail discussion with FPI team (mainly Stephen) and Mark,
latest e-mail thread dated 2020-01-23T16:50 CET. The FPI team would
really like to have our L2 scpot for their calibration regardless of its
poor state when running in eclipses (with all elements biases at 0 nA).

A compromise was reached in that the last run of Mark's cleanup script
(typically 36 hours after) not applying the masking. (This will also be
the case for any later re-processing). This way the initial L2 scpot
products will have the eclipse and maneuvers set to FillValues and SITL
not seeing them (as before) but the products in the end having all
available data.