MMS db start work looking up Pred.Eph. as well
this might be required to process new L2a dce2d if NO Def.Eph. is
provided as argument and NO DefEph could be located by the script..

These files however is somewhat more tricky to combine as there are some
files which cover very long time periods (with low time resolution) and
some other files with short durations overlapping (with higher time
resolution). The commit here simply lists which files might be of
interest based on the time interval requested... Using their file name
as a first filtering in order to reduce the I/O reading required (to
speed up the locating of files).