MMS use lookup of SCPOT file for SLOW L2a->L2pre
Moving lookup outside of mms_sdp_get_offset (to ensure it can be stored
as a proper CDF GlobalAttrib "Parent" in the created Slow L2pre dce2d
file to ensure we keep full tracability of input files used! (NASA
requirement for MMS). If no SCPOT file is given as input at SDC then go
looking for it to match the L2a file, as we are only interested in one
CDF parameter we simply store it alone as a TSeries in the field
"scpotFile".