MMS include subtracted DSL offset in L2Pre dce2d
Following in person discussion with @danbgraham, DSL offsets should be
included in the created L2Pre dce2d files. This as the new scpot
dependency (SLOW mode) is not as straight forward as the old offset
indicated by the GlobalAttribute "Calibration_file". It was also a
request made by LASP to include this in our files, it should however not
be necessary to reprocess files other than the SLOW mode (starting with
L2A) for the entire mission.

This change is breaking backwards/forwards compatibility (a new cdf
parameter) and as such increment file version X number for L2Pre files.