MMS fallback to AFG if DFG is missing for L2pre
As noted in e-mail from Mark dated 2020-04-18T16:54 CEST, the MMS2 DFG
was switched off for a period of time that week and as a result NO DFG
files will be produced! Our dependency on DFG would therefor limit our
own products, this commit will try to allow for a fallback to AFG
(given as inputs), if no DFG files are present.
It should be noted that if the switch is done, it should not mix and
match AFG and DFG files for example for burst segements. It should use
one or the other of the fluxgates for each day or burst segement.