Create CODEOWNERS
Adding a new CODEOWNERS file, detailed description: https://help.github.com/en/github/creating-cloning-and-archiving-repositories/about-code-owners.

This should hopefully help in assigning code reviewer(-s) whenever an issue or pull request is opened for irfu-matlab.

NOTE: If this works then it should perhaps be refined with some more rules and patterns in the future. Also note, for now, I do not include some major contributors to irfu-matlab (like Yuri and Andris as they are very busy as is). It should still be possible to manually notify and assigned people to issues other than the defined "codeowners", as needed!