Make use of common time functions in get_position
Dear @Niked82, welcome to irfu-matlab! I took the liberty to slightly
change your function so that it makes use of some of our time rutines
here in irfu-matlab. Hope you do not mind.

The reason of this is mainly to avoid problems with what time it
realy is* and as some (if not all) of SolO CDF files already use
CDF_EPOCH_TT2000 the code to get positon could make use of the same
functions for converting time (and convert time only if needed).

* If you want to go down that rabbit hole, look up details such as
"The Relationship between TDT and TDB", "leapseconds", etc. I do not
think Matlab's datestr() handle such things...