Use only newer "irf.log()" for the newer missions
Related to the old issue number 9, it is better if newer missions like
mms and solarorbiter use the new irf.log(). Older missions like Cluster
however will for the forseable future keep using the older "irf_log()"
function (as it has a lot of other scripts which depend on irf_log() in
one way or another).