[WIP] a draft version of "get_mkernel"
Keep metakernel paths stored between sessions, convert paths to kernel
files so that they work on independently of computer (local or IRFU
servers). If the old kernel path is not present/defined, then simply ask
the user for input.

Related to the old issue #6.

Note: Windows is not supported. And marked as "Work-In-Progress" in a
separate branch as it is not yet tested!