Add a try & catch statement to irf('ceflib')
allowing for a failure of loading the compiled CEF library without a
full crash. And improve messages to users (only if the user needs the
required libs for their inteded use of irfu-matlab should they be
requested to contact IRFU). CEF lib is used for Cluster Exchange Files,
which is typically not used for other missions, IRBEM is typically used
for coordinate transformations and MICE/SPICE is of geometry
information etc. of various space missions.

Related to issue #66.