[CI] Create .travis.yml
Add a first test version of .travis.yml,
Note: typically changes like this should be merged to "devel", not directly to "master". But according to my understanding of the CI/CD documentation it must be present in the default branch for TravisCI to detect it and use it.
Related to commit ce08ba2cc8e5908325d9db385df77f3dfa021e9f.