Adding a more proper ref for phase angles
as I was looking of this and started wondering from where it came...
E-mail from PAL forwarding I.Dors e-mail from 2014-04-11 contained a
statement:
"Your 29.85-deg angle is correct. The 33-deg is very, very old
information and the diagram is in need of an update."

The "diagram" in question is a "Fig 13 (Electric Sensor System) in a
FIELDS instrument paper".