Style change Fortran<=72 char per line and no tabs
Standard Fortran 90 does not allow tabs (nor older does older Fortran),
but some compilers simply "translate" it anyhow. This change makes the
code more in-line with offical standards. And I have applied two space
indentations in do & if statements and similar.

According to older Fortran standards:
The column 1 is used for comments ("C"), columns 1-5 are otherwise
typically used for line numbers (go-to statements), column 6 can be
used for continuations ("+" in column 6 means it is treated as
following the previous line) and column 7 is the default starting
position for the code.

Note: the commit here looks large but it is mostly whitespace changes,
which can be verified with the following command:
`git diff --ignore-space-change c_efw_spinfit_mx.F`.

With this commit it should no longer be required to compile the mex on
Linux with the special "-ffixed-line-length-none" argument.