Start work on Cluster CSA move from CAIO to TAP
IMPORTANT NOTE: The following CAIO features are not yet available in
TAP (according to https://www.cosmos.esa.int/web/csa/caiototap):
* Asynchronous product requests (data)
* Header requests (counts as data)
* Streaming data requests (data)

Note: The metadata and inventory requests appears to be drastically
changed meaning this commit here is broken (as it depends on listing the
available data). More commits will follow.

Related to issue #75, kept in a separate branch until ready for merge to
devel.