[minor] make use of HeaderLines (untested)
Try to fix last commit (16b546483de12b740fad858117713b79e0f48c12) for
other returnTimeTable values. Note this is untested.

CSA TAP metadata requests have changed a lot more, so now one must ask
for metadata from a specific (sql?) table. For the basic "inventory" it
is part of query "select ... FROM+csa.v_dataset_inventory+", but it is
not obvious what the new table names are...
For instance one of the examples on the transition website
https://www.cosmos.esa.int/web/csa/caiototap uses `FROM+csa.v_dataset`.

Related to issue #75.