MMS edp sweep keep track method used for Impedance
and give a warning if replacing it. Otherwise, it is possible to first
run "analyse()" or "analyse_all()" which computes ".impedance" according
to one method and then run "plot_time()" which replaces the ".impedance"
with value computed with another method. If then user calls the basic
"plot()" function the value shown as "impedance" is not the one from
"analyse()" but from "plot_time()" while it shows as "impedance @
current value".

This can lead to some confusion, so now just give a warning to the user
whenever impedance is replaced with value of the other method...