Use webread & catch ME if problem checking version
When trying to read upstream version a message of "Not connected to
internet" is not very helpful to debug what exactly went wrong.
It's better to write the corresponding error message to see what is
actually going on. And try to make use of the recommended webread()
instead of the legacy urlread(), if running newer Matlab versions.

Per-Arne experienced some problems with fetching upstream version
information despite being connected to the internet. While this does not
fully solve the problem it should perhaps help debugging it...