Minor clean up of solo/get_data
Make use of "switch & case" instead of sequential if-statements which
are executed as if they could be combined (first if statement fully ends
then next if statement is checked..). Having 'switch & case' or just
basic 'if, elseif, else, end' makes the branching of the executed code
much clearer to any human reader.