MMS sweep analysis code update, compute "iPh knee"
Details discussed in e-mails and teleconfs back in 2021-09 between
Per-Arne, Andris, Yuri, Daniel G. and me.

Long story short:
1) code now tries to split "-+" and "-+" type sweeps and compute two values
each per sweep for these (interpreting "-" and "+" parts separately).
Previously these sweeps were simply not computed.

2) try to compute a "knee" intersection of the almost vertical and
almost horizontal part of the sweep based on the spread in the resulting
voltage measured (vertical part have very little spread in measured
voltage per bias current setting while the horizontal part of the sweep
have a very large spread). ("Vertical/Horizontal" when plotting voltage
responses at X-axis and bias current settings at Y-axis). If no "knee"
is found this value resort back to being simply the old "Iph" value,
that is the bias current value at "(2*Vmin+V0)/3".

3) Phase output is now the z-phase of this "knee" intersection.

The reason for trying to find this "knee" is that it is a (current)
region in which the probe I-V response is no longer simply linear and we
want to avoid this region when commanding the bias currents of the
probes. Similar to "iPh", which is typically computed as a value
slightly higher (20-30nA) than this "knee" intersection, especially when
running sweeps of type "--".