Mutually excl. subsequent if's => one switch
Having mutually exclusive if statements test one single parameter
(ignoring case) result in 10 or so extra checks if the first entry
matched as the if-statement fully ended.
This change here replace it all with one single switch statement.

Also add a note for "PAS ion VDFs", which must be re-written as it use
a lot of hard coded paths and converting times back and forth from
TT2000 (aka 'epochtt') for no apparent reason...