MMS sweeps include commanded DAC currents in plots
Based on the extracted TSeries of the commanded DAC bias currents
(commit 4fc73d68) the iPh and iPh_knee plots now show the historical
operating point used (and by comparing the DAC and iPh_knee it is clear
that sometimes we have not had that much in terms of margins).

Note: The DAC values plotted are the negative values commanded (due to
different convention of current direction used between the DAC and iPh,
e.g. flowing in [or out] of the probe from [or to] the plasma).