Two minor bugfixes in caa_download
1. `downloadedFile` did not have any default value. This caused trouble
because the new TAP interface return 404 error on some requests which
is handled by the `try,catch` but when then when this second output
argument is missing it results in new error not always handled correctly
upstream in the code.

2. the split'ed `tmpGetRequest` cell does not always have 10 or 12
elements. It is better to simply print all of the existing elements to
the log than use hard-coded positions (the main data product requested
can have different positons depending on exactly how it was requested).

These two bugs where found earlier this week by @JanKarlsson when
testing the new ESA Cluster TAP interface as part of solving issue #75,
which must now be completed withing less than a month.