MMS3 sdp 2 failed as of 2022-05-08T23:18:44.300
(or at least it appears to have failed, investigation is ongoing).
It appears to have been another micrometeorite hit, this time for MMS3
taking out our SDP 2. Looking at sweeps following this it does not
appear to be much I-V response to speak of, other instruments reported a
drastic change in S/C potential at this time and the spin rate had a
drastic jump...

This commits adds initial support for reconstructing the MMS3 E12 data
from p134, similar to how it was done for the MMS2 sdp2 failure.

However, please note that the hardcoded values of 'cmdAv' in
mms_sdp_model_spin_residual_cmd312.m was simply copied from MMS2 sdp2,
they should be changed to use proper values for MMS3.