Fix order of latched & reconstructed
Previously latched probes (stuck at a fixed value) were identified
before the code checked for failed probes. When probes have failed they
are typically mostly giving bad values, meaning a false positive for
"stuck probes" was detected. However, when we reconstruct the failed
probe later in the process its value is no longer "stuck".

This false postive for stuck probes results in a lower quality reported
for much of the data after the respective spacecraft had its first
failure. (Not constant, but much of the data has this issue). This was
first noticed and reported by Narges @ LASP.