MMS warn when sweep occur during eclipse/maneuver
Some analyzed sweep result tend to look like outliers when comparing against
a timeseries of other sweeps results. After looking into the mission
timeline xml file (eg. mms_timeline_2023191_2023265_v02.xml) I found one
possible explaination for this, namely it was doing a sweep during a
time when spacecraft was in eclipse caused by the Earth.

Specifically I have noted that mms3_edp_srvy_l1b_sweeps_20230714_v0.7.4.cdf,
sweep nr 10, nr 11 and nr 12, (executed during the interval from
2023-07-14T14:43:21.492053610Z to 2023-07-14T14:43:38.492294779Z), where
one instance where sweeps coincided with eclipse.

The reason why the sweep results are so very different is because of the
lack of photoelectrons, this very different environment should be noted
and sweep results should be used with caution (i.e. do not rely on these
sweeps for determining which bias should be used when spacecraft is not
in eclipse).

Also included in this commit is a switch over to rely more on Matlab's
build in argument parser, which validate some of the inputs.