Add some .pre-commit checks, run before git commit
As sometimes misstakes are made (such as missing a few "<< == >>" when a
merge conflict occurs) it is better to try and catch these early on and
prevent the commits from being made locally in the first place.

(Note: pre-commits can be bypassed if needed, and must be installed in
the first place in order to run each time one is about to make a
commit. But they can be useful in order to catch some misstakes, here I
am adding only some out-of-the-box hooks and nothing specific to
Matlab).

Also try to move towards the new GitHub issue template form (allowing
for required checkboxes and a bit more structured form). This will not
be visible until it is in master branch.
Ref: https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/syntax-for-issue-forms