irf_colormap 'batlow' with proper ref. and LICENSE
Values are now as per 'batlow.txt', in v8.0.1,
DOI: 10.5281/zenodo.8409685.

Also included here is the LICENSE text (MIT), a requirement for us to be
compliant with the selected LICENSE of the work ('batlow'), specifically
the line:
"The above copyright notice and this permission notice shall be included
in all copies or substantial portions of the Software."

(also included is a corresponding "SPDX-License-Identifier:"-comment, as
this can help genarate a "software bill of materials" and help licensing
issues in the future).