A few more Matlab lint fixes
- Use "system(['X'])" instead of "eval(['!X'])".
- Avoid unneccessary "[ ]" around single scalar values.
- Use "isempty(X)" instead of "length(X)>0".
- And a few more uncreachable & comma in if-statements.

And correct an MMS "if all(diff(enabled.data))==0", which was, as far as
I can tell and recall to be "if all(diff(enabled.data)==0)".

(If probes are enabled or all disabled, then diff of status is zero, if
all diff status is zero that means it's not changing on/off which the thing
we wanted to detect and potentially give error on).