Update LIBIRBEM with IGRF-14 coef. (Win&Linux)
Compiled on Ubuntu 24.04 with gfortran v13.3.0 for Linux.

Compiled on Windows 10 with gfortran v13.2.0 (MSYS2 UCRT64).

Note; I have been unable to locate a Mac OSX machine with a working
gfortran install which is why the corresponding file for Mac is still
the old one.

Source of libirbem used here is the github repo PRBEM/IRBEM at commit
8279aa9fd50c0a32d63efe762454b9ebf3b81ef6.