CI GitHub runners ubuntu-24.04 & ubuntu-latest
As the currently used ubuntu-20.04 is to become deprecated next year,
(ref: github.com repo actions/runner-images issue 11101), we need to
move to newer GitHub runner images.

The "ubuntu-24.04" should have libgfortran5 available for install
(it is available on my installed ubuntu 24.04 machine), but the SOdevel
specific tests do not rely on the compiled fortran code and should be
able to run on the moving target "ubuntu-latest".