MMS Workaround for nonmonotonically increasing DFG
As was the case for L2pre dce2d processing of MMS3 2023-05-18 when the
currently latest version of the DFG L2pre file does have some duplicate
timestamps (mms3_dfg_srvy_l2pre_20230518_v5.399.0.cdf).

Here I have implemented the same approach as we do use for DefAtt (a
product which is expected to have duplicates sometimes), that is to say
we discard the first datapoint of the duplicate and use only the last
one in the file. After the initial clean up a check for
non-monotonically increasing times is still applied (our cleanup does
not touch negative times only duplicated timestamps).