Add warning in "spin residual" if phase doubles up
That is to say, if phase was computed from a HK file with several large
gaps it can be computed in segements and interpolated. But if it during
the gaps change spin revolution rate then the rate change may introduce
some negative jumps backward in the unwrapped (2*pi) continuous phase.

This was the case for mms4_fields_hk_l1b_101_20250421_v0.5.1.cdf, and as
such the "timeTmp = interp1(phaDegUnw,epochTmp,phaFixed);" fails.

This commit is just adding a warning to the logs, perhaps to solve it
ultimately one should also compute the spin residuals in segments just
as the "phase" is computed by finding continous segments..
Or perhaps it is sufficient to just wait for the next, hopefully better
HK 101 (or DefAtt) file which has all the necessary spin phase
information properly filled in from the start.