[minor] re-apply irf_indentations.m aligning code
As is noted in our CONTRIBUTING guidelines code should be written with
two (and only 2) spaces indentation (and only spaces, not tabs).

Not applying the indentation rule on any SolO code (not sure if
currently even needed) following in person request from Erik J as to
avoid potential merge conflicts.