Matlab R2025a+ may have incompatibility issues
As found in GenericTimeArray problem described in issue 169, some
code in irfu-matlab relies on the old behaviour of `:` (colon) in
Matlab. This was changed in R2025a and may result in errors if any of
the values being executed are not scalar.

As Matlab writes:
> `colon` now returns an error when creating vectors if one or more
> operands are not scalar. For example, expressions like
> `[1 2 3]:2:10` now error. Previously, `colon` used the first element
> of any nonscalar operands to evaluate the expression.

Ref. https://se.mathworks.com/help/matlab/ref/double.colon.html