Fix more potential R2025a+ colon problems
That is to say "idx=find(foobar)" could return multiple values, if then
later trying to use the the "idx" in some array construction i.e.
"1:ind" then R2025a+ would result in an error. The old behaviour of
using the first match can be achived by simply telling "find" to only
find 1 match (and take the 'first', or 'last' as per code use).

Also added two clearifying comments in mms_region_times on how it treats
the indicators (not as "logical and" => require all to be true but rather
"adding" the logical arrays togeather and checking if sufficiently many
of them indicate one it is SW).