BICAS: 3xBugfix: vdccal(): Correct
(1) Derive PSP using only ANT1 (without calibration data) when there is no
calibration data.

(2) Correct criterion for ANT2 or ANT3 not being available for deriving PSP
(OR instead of AND).

(3) Not use calibration data (all data products) outside of time interval
for which it is derived (set calibration data to NaN).